backend/stable-diffusion/requirements.txt
stability-sdk
Pillow